backend/requirements.txt
fastapi
uvicorn
python-multipart
docxtpl
python-docx
